Edit to `Intro_to_R` exercise notebook instructions
In line 121, the metadata to be stored into the variable in this step of the intro_to_R_tidyverse_exercise notebook was clarified to be the renamed metadata from the previous step. The line, "Hint: `metadata_clean` should be placed before your first pipe operator" was also added to steer participants away from using the original metadata.