Updates to include line numbers
Note that this change requires moving to `html_document` output, so we now have `*.html` files in addition to `*.nb.html` files.  
I have updated links to point to the bare `.html` files, but this is a change that may have effects outside this repository.

Finally, as this now requires the full page to knit in one go, I have added some more `eval=GALSE` chunks as well as an `include=FALSE` where user interaction was required to create a data frame. This is done well below where it is initially asked for, so as a pedagogical matter, even students working from the notebook are unlikely to get a "hint".